==> Machine Learning/FEML_Codes/01.1. CreditApprovalUCI_dataPrep.ipynb <==
Prepare the **Credit Approval data set**, to leave it more suitable for the demos of the recipes from chapter 3.

import random
import pandas as pd
import numpy as np

data = pd.read_csv('crx.csv')
data.head()

data.shape 

for col in data.columns:
    print(col, type(col))

**Create variable names according to UCI Machine Learning! Re-cast some variables to the correct types!**

varnames = ['A'+str(s) for s in range(1,17)]
data.columns = varnames
data = data.replace('?', np.nan)  # replace ? by np.nan
data['A2'] = data['A2'].astype('float')
data['A14'] = data['A14'].astype('float')
data['A16'] = data['A16'].map({'+':1, '-':0})  # encode target to binary
data.head()

**Add more missing values to random positions! I will help with the demos of the recipes!**

random.seed(9001)
values = set([random.randint(0, len(data)) for p in range(0, 100)])
for var in ['A3', 'A8', 'A9', 'A10']:
    data.loc[values, var] = np.nan
data.isnull().sum()

**Save the data!**

data.to_csv('creditApprovalUCI.csv', index=False)

data.head()

cat_cols = [c for c in data.columns if data[c].dtypes=='O']
data[cat_cols].head()  # categorical variables

num_cols = [c for c in data.columns if data[c].dtypes!='O']
data[num_cols].head()  # numerical variables
