==> Machine Learning/FEML_Codes/01.2. Titanic_dataPrep.ipynb <==
## Predicting Survival on the Titanic

import pandas as pd
import numpy as np
import seaborn as sns

data = pd.read_csv('titanic.csv')
data.head()

data.shape

data.describe()

data = data.replace('?', np.nan)
data.isnull().sum()

def get_first_cabin(row):
    try:
        return row.split()[0]
    except:
        return np.nan 

data['cabin'] = data['cabin'].apply(get_first_cabin)

data.to_csv('../titanic.csv', index=False)
