==> Machine Learning/FEML_Codes/03.1-Missing-Data.ipynb <==
## Missing Values

### Missing Data Mechanisms

**1. MCAR** Missing Completely at Random (Same probability, no bias) **2. MAR:** Missing at Random, if men are more likely to disclose their weight than women, **3. Missing Not at Random, MNAR:** if people failed to fill in a **depression survey** because of their level of depression.

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
pd.set_option('display.max_columns', None)  # to display the total number columns in the dataset  

data = pd.read_csv('titanic.csv')
data.head()

In python, the missing values are stored as NaN.

data.isnull().sum()

There are **263 missing values for Age, 1014 for Cabin and 2 for Embarked**.

data.isnull().mean()

There are missing data in the variables **Age (20% missing), Cabin (77% missing), and Embarked** -the port from which the passenger got into the Titanic- **(~0.2%  missing)**.

### Missing data Not At Random (MNAR): Systematic missing values

In the Titanic dataset, we could expect a greater number of missing values in **age** and **cabin** for people who did not survive. Because the people who survived could be otherwise asked for that information.

data['cabin_null'] = np.where(data['cabin'].isnull(), 1, 0)

data.groupby(['survived'])['cabin_null'].mean()

data['cabin'].isnull().groupby(data['survived']).mean()  # another way for the above!

The percentage of missing values is higher for people **who did not surviv**e (87%), respect to people **who survived** (60%).

data['age_null'] = np.where(data['age'].isnull(), 1, 0)
data.groupby(['survived'])['age_null'].mean()

data['age'].isnull().groupby(data['survived']).mean()

There is a higher number of missing data for the people who did not survive the tragedy.

### Missing data Completely At Random (MCAR)

data[data['embarked'].isnull()]

These 2 women were traveling together, Miss Icard was the maid of Mrs Stone.

### Missing data at Random (MAR)

In the ficticious **loan book toy dataset** example, data missing in employment are associated with data missing in time in employment.

data = pd.read_csv('loan.csv', usecols=['employment', 'time_employed'])
data.head()

data.isnull().mean()  # Check the percentage of missing data

We see that both variables have the **same percentage** of missing observations roughly.

data['employment'].unique()  # Examples of employments

print('Number of employments: {}'.format(    # Their numbers
    len(data['employment'].unique())))

We observe the **missing information (nan)** and several different employments of the people.

data['time_employed'].unique()  # The variable time employed

The customer can't enter a value for **employment time** if they are **not employed**. They could be **students, retired, self-employed, or work in the house**. But we see how these 2 variables are related to each other.

**percentage of missing data in time employed for the customers who declared employment**!

t = data[~data['employment'].isnull()]  # customers who declared employment
t['time_employed'].isnull().mean()  # percentage of missing data in time employed

**Missing data persentage in the customers who did not declare employment!**

t = data[data['employment'].isnull()]
t['time_employed'].isnull().mean()

The number of borrowers who have reported occupation and have missing values in time_employed is minimal. Whereas the customers who did not report an occupation or employment, are mostly reporting missing values in the time_employed variable
