==> Machine Learning/FEML_Codes/03.2-Cardinality.ipynb <==
## Cardinality

The values of a categorical variable are categories (labels ) can contains only 2 labels like gender, or more like city or postcode. A **high number of labels** within a variable is known as __high cardinality__. We should reduce the number of labels to boost performance in machine learning.

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
from sklearn.linear_model import LogisticRegression  # Build ML models
from sklearn.ensemble import AdaBoostClassifier
from sklearn.ensemble import RandomForestClassifier
from sklearn.ensemble import GradientBoostingClassifier
from sklearn.metrics import roc_auc_score  # Evaluate the models
from sklearn.model_selection import train_test_split  # Separate data into train and test

data = pd.read_csv('titanic.csv')
data.head()

The categorical variables in this dataset are **Name, Sex, Ticket, Cabin and Embarked**. **Ticket and Cabin** contain both letters and numbers, so they could be treated as Mixed Variables. Now, we will treat them as categorical.

**How is the cardinality of these categorical variables?**

print('Number of categories in the variable Name: {}'.format(
    len(data.name.unique())))
print('Number of categories in the variable Gender: {}'.format(
    len(data.sex.unique())))
print('Number of categories in the variable Ticket: {}'.format(
    len(data.ticket.unique())))
print('Number of categories in the variable Cabin: {}'.format(
    len(data.cabin.unique())))
print('Number of categories in the variable Embarked: {}'.format(
    len(data.embarked.unique())))
print('Total number of passengers in the Titanic: {}'.format(len(data)))

While the variable **Sex** contains only 2 categories and **embarked** 4 (low cardinality), the variables **Ticket, Name and Cabin** has high cardinality.

data.cabin.unique()

We capture only **the first letter** that indicates the deck on which the cabin was located, and therefore indicates social class status and proximity to the surface of the Titanic.

data['Cabin_reduced'] = data['cabin'].astype(str).str[0]
data[['cabin', 'Cabin_reduced']].head()

print('Number of categories in the variable Cabin: {}'.format(
    len(data.cabin.unique())))
print('Number of categories in the variable Cabin reduced: {}'.format(
    len(data.Cabin_reduced.unique())))

We reduced the number of different labels **from 182 to 9**.

**Separate into training and testing set for ML models!**

use_cols = ['cabin', 'Cabin_reduced', 'sex']
X_train, X_test, y_train, y_test = train_test_split(
    data[use_cols], 
    data['survived'],  
    test_size=0.3,
    random_state=0)
X_train.shape, X_test.shape

### High cardinality leads to uneven distribution of categories in train and test sets

unique_to_train_set = [
    x for x in X_train.cabin.unique() if x not in X_test.cabin.unique()]
len(unique_to_train_set)

There are **113 Cabins** only present in the training set, and **not in the testing set**.

unique_to_test_set = [
    x for x in X_test.cabin.unique() if x not in X_train.cabin.unique()]
len(unique_to_test_set)

This **over-fitting problem** or the problem not to know how to handle is almost overcome by **reducing the cardinality** of the variable. See below.

unique_to_train_set = [
    x for x in X_train['Cabin_reduced'].unique()
    if x not in X_test['Cabin_reduced'].unique()]
len(unique_to_train_set)

unique_to_test_set = [
    x for x in X_test['Cabin_reduced'].unique()
    if x not in X_train['Cabin_reduced'].unique()]
len(unique_to_test_set)

Observe how by reducing the cardinality there is now only 1 label in the training set that is not present in the test set. And no label in the test set that is not contained in the training set as well.

### Effect of cardinality on Machine Learning Model Performance

In order to evaluate the effect of categorical variables in machine learning models, I will quickly replace the categories by numbers. See below.

**Re-map Cabin into numbers so we can use it to train ML models!**

cabin_dict = {k: i for i, k in enumerate(X_train.cabin.unique(), 0)}
cabin_dict

X_train.loc[:, 'Cabin_mapped'] = X_train.loc[:, 'cabin'].map(cabin_dict)
X_test.loc[:, 'Cabin_mapped'] = X_test.loc[:, 'cabin'].map(cabin_dict)
X_train[['Cabin_mapped', 'cabin']].head(10)

**Replace the letters in the reduced cabin variable with the same procedure! First create replace dictionary!**

cabin_dict = {k: i for i, k in enumerate(X_train['Cabin_reduced'].unique(), 0)}
X_train.loc[:, 'Cabin_reduced'] = X_train.loc[:, 'Cabin_reduced'].map(cabin_dict)
X_test.loc[:, 'Cabin_reduced'] = X_test.loc[:, 'Cabin_reduced'].map(cabin_dict)
X_train[['Cabin_reduced', 'cabin']].head(20)

We see now that E36 and E24 take the same number, 1, because we are capturing only the letter. They both start with E.

**Re-map the categorical variable Sex into numbers!**

X_train.loc[:, 'sex'] = X_train.loc[:, 'sex'].map({'male': 0, 'female': 1})
X_test.loc[:, 'sex'] = X_test.loc[:, 'sex'].map({'male': 0, 'female': 1})
X_train.sex.head()

**Check the missing values!..**

X_train[['Cabin_mapped', 'Cabin_reduced', 'sex']].isnull().sum()

X_test[['Cabin_mapped', 'Cabin_reduced', 'sex']].isnull().sum()

In the test set, there are now 41 missing values for the highly cardinal variable. These were introduced when encoding the categories into numbers. 

How? 

Many categories exist only in the test set. Thus, when we created our encoding dictionary using only the train set, we did not generate a number to replace those labels present only in the test set. As a consequence, they were encoded as NaN. We will see in future notebooks how to tackle this problem. For now, I will fill those missing values with 0.

**Check the number of different categories in the encoded variables!**

len(X_train.Cabin_mapped.unique()), len(X_train.Cabin_reduced.unique())

From the above we note immediately that from the **original 182 cabins** in the dataset, **only 147 are present in the training set**. We also see how we **reduced the number of different categories to just 9** in our previous step.

### Random Forests

**Build model on data with high cardinality for cabin! Call the model!..**

rf = RandomForestClassifier(n_estimators=200, random_state=39)
rf.fit(X_train[['Cabin_mapped', 'sex']], y_train)
pred_train = rf.predict_proba(X_train[['Cabin_mapped', 'sex']])
pred_test = rf.predict_proba(X_test[['Cabin_mapped', 'sex']].fillna(0))
print('Train set')
print('Random Forests roc-auc: {}'.format(roc_auc_score(y_train, pred_train[:,1])))
print('Test set')
print('Random Forests roc-auc: {}'.format(roc_auc_score(y_test, pred_test[:,1])))

rf = RandomForestClassifier(n_estimators=200, random_state=39)
rf.fit(X_train[['Cabin_reduced', 'sex']], y_train)  # train the model
pred_train = rf.predict_proba(X_train[['Cabin_reduced', 'sex']])  # make predictions
pred_test = rf.predict_proba(X_test[['Cabin_reduced', 'sex']])
print('Train set')
print('Random Forests roc-auc: {}'.format(roc_auc_score(y_train, pred_train[:,1])))
print('Test set')
print('Random Forests roc-auc: {}'.format(roc_auc_score(y_test, pred_test[:,1])))

We observe that **the performance of the Random Forests on the training set** is quite superior to its performance in the test set. This indicates that the model is **over-fitting**, which means that it does **a great job at predicting the outcome on the dataset** it was trained on, but **it lacks the power to generalise the prediction** to unseen data.

**Now build  again model on data with low cardinality for cabin! Call the model!..** We can see now that **the Random Forests no longer over-fit to the training set**. In addition, the model is much better at generalising the predictions (compare the roc-auc of this model on the test set vs the roc-auc of the model above also in the test set: 0.81 vs 0.80).

**We can overcome the effect of high cardinality by adjusting the hyper-parameters of the random forests**. Here, I want to show you that given a same model, with identical hyper-parameters, **high cardinality may cause the model to over-fit**.

### AdaBoost



**Build model on data with plenty of categories in Cabin! Call the model!..**

ada = AdaBoostClassifier(n_estimators=200, random_state=44)
ada.fit(X_train[['Cabin_mapped', 'sex']], y_train)# train the model
pred_train = ada.predict_proba(X_train[['Cabin_mapped', 'sex']])  # make predictions
pred_test = ada.predict_proba(X_test[['Cabin_mapped', 'sex']].fillna(0))
print('Train set')
print('Adaboost roc-auc: {}'.format(roc_auc_score(y_train, pred_train[:,1])))
print('Test set')
print('Adaboost roc-auc: {}'.format(roc_auc_score(y_test, pred_test[:,1])))

**Build model on data with fewer categories in Cabin! Call the model!..**

ada = AdaBoostClassifier(n_estimators=200, random_state=44)
ada.fit(X_train[['Cabin_reduced', 'sex']], y_train)  # train the model
pred_train = ada.predict_proba(X_train[['Cabin_reduced', 'sex']])  # make predictions
pred_test = ada.predict_proba(X_test[['Cabin_reduced', 'sex']].fillna(0))
print('Train set')
print('Adaboost roc-auc: {}'.format(roc_auc_score(y_train, pred_train[:,1])))
print('Test set')
print('Adaboost roc-auc: {}'.format(roc_auc_score(y_test, pred_test[:,1])))

Similarly, the Adaboost model trained on the variable with **high cardinality is overfit** to the train set. Whereas the Adaboost trained on the **low cardinal variable is not overfitting** and therefore does a better job in generalising the predictions.

In addition, building an AdaBoost on a model with less categories in Cabin, is a) **simpler** and b) should a **different category** in the test set appear, by taking just the front letter of cabin, the ML model will know how to handle it because it was seen during training.

### Logistic Regression

**Build model on data with plenty of categories in Cabin! Call the model!..**

logit = LogisticRegression(random_state=44, solver='lbfgs')
logit.fit(X_train[['Cabin_mapped', 'sex']], y_train)  # train the model
pred_train = logit.predict_proba(X_train[['Cabin_mapped', 'sex']])  # make predictions
pred_test = logit.predict_proba(X_test[['Cabin_mapped', 'sex']].fillna(0))
print('Train set')
print('Logistic regression roc-auc: {}'.format(roc_auc_score(y_train, pred_train[:,1])))
print('Test set')
print('Logistic regression roc-auc: {}'.format(roc_auc_score(y_test, pred_test[:,1])))

**Build model on data with fewer categories in Cabin! Call the model!..**

logit = LogisticRegression(random_state=44, solver='lbfgs')
logit.fit(X_train[['Cabin_reduced', 'sex']], y_train)  # train the model
pred_train = logit.predict_proba(X_train[['Cabin_reduced', 'sex']])  # make predictions
pred_test = logit.predict_proba(X_test[['Cabin_reduced', 'sex']].fillna(0))
print('Train set')
print('Logistic regression roc-auc: {}'.format(roc_auc_score(y_train, pred_train[:,1])))
print('Test set')
print('Logistic regression roc-auc: {}'.format(roc_auc_score(y_test, pred_test[:,1])))

We can draw the same conclusion for Logistic Regression: **reducing the cardinality improves the performance and generalisation of the algorithm**.

### Gradient Boosted Classifier

**Build model on data with plenty of categories in Cabin! Call the model!..**

gbc = GradientBoostingClassifier(n_estimators=300, random_state=44)
gbc.fit(X_train[['Cabin_mapped', 'sex']], y_train)  # train the model
pred_train = gbc.predict_proba(X_train[['Cabin_mapped', 'sex']])  # make predictions
pred_test = gbc.predict_proba(X_test[['Cabin_mapped', 'sex']].fillna(0))
print('Train set')
print('Gradient Boosted Trees roc-auc: {}'.format(roc_auc_score(y_train, pred_train[:,1])))
print('Test set')
print('Gradient Boosted Trees roc-auc: {}'.format(roc_auc_score(y_test, pred_test[:,1])))

**Build model on data with fewer categories in Cabin! Call the model!..**

gbc = GradientBoostingClassifier(n_estimators=300, random_state=44)
gbc.fit(X_train[['Cabin_reduced', 'sex']], y_train)  # train the model
pred_train = gbc.predict_proba(X_train[['Cabin_reduced', 'sex']])  # make predictions
pred_test = gbc.predict_proba(X_test[['Cabin_reduced', 'sex']].fillna(0))
print('Train set')
print('Gradient Boosted Trees roc-auc: {}'.format(roc_auc_score(y_train, pred_train[:,1])))
print('Test set')
print('Gradient Boosted Trees roc-auc: {}'.format(roc_auc_score(y_test, pred_test[:,1])))

**Gradient Boosted trees are indeed over-fitting** to the training set in those cases where the variable Cabin has a lot of labels. This was expected as **tree methods** tend to be biased to variables with plenty of categories.
