==> Machine Learning/FEML_Codes/04.08-Missing-Indicator.ipynb <==
## Adding a variable to capture NA

**Mean / median and random sample imputation assume that the data are (MCAR)**. **Arbitrary value imputation or end of distribution imputation techniques will affect the variable distribution dramatically**, and are therefore not suitable for linear models. If data are **not missing at random (MNAR)**, it is a good idea to **replace missing observations by the mean / median / mode** AND  **flag** those missing observations as well with a **Missing Indicator**. A Missing Indicator is an **additional binary variable**, which indicates whether the data was missing for an observation (1) or not (0). We can add a missing indicator to **both numerical and categorical variables**. It is easy to implement and captures the importance of missing data if there is one!

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
from sklearn.model_selection import train_test_split

data = pd.read_csv('titanic.csv', usecols=['age', 'fare', 'survived'])
data.head()

data.isnull().mean()

To add a **binary missing indicator**, we don't necessarily need to learn anything from the training set, so in principle **we could do this in the original dataset** and **then separate into train and test**. But it is not preferable! Now, let's create a **binary missing indicator manually**.

X_train, X_test, y_train, y_test = train_test_split(
    data[['age', 'fare']],  # predictors
    data['survived'],  # target
    test_size=0.3,  # percentage of obs in test set
    random_state=0)  # seed to ensure reproducibility
X_train.shape, X_test.shape

X_train.isnull().mean()

X_train['Age_NA'] = np.where(X_train['age'].isnull(), 1, 0)
X_test['Age_NA'] = np.where(X_test['age'].isnull(), 1, 0)
X_train.head()

X_train['Age_NA'].mean()

X_train.isnull().mean()

median = X_train['age'].median()
X_train['age'] = X_train['age'].fillna(median)
X_test['age'] = X_test['age'].fillna(median)
X_train.isnull().mean()

### House Prices dataset

cols_to_use = ['LotFrontage', 'MasVnrArea', # numerical
               'BsmtQual', 'FireplaceQu', # categorical
               'SalePrice' ] # target

data = pd.read_csv('HousingPrices_train.csv', usecols=cols_to_use)
print(data.shape)
data.head()

data.isnull().mean()

X_train, X_test, y_train, y_test = train_test_split(data,
                                                    data['SalePrice'],
                                                    test_size=0.3,
                                                    random_state=0)
X_train.shape, X_test.shape

def missing_indicator(df, variable):    
    return np.where(df[variable].isnull(), 1, 0)

**Loop over all the variables and add a binary missing indicator with the function we created!**

for variable in cols_to_use:
    X_train[variable+'_NA'] = missing_indicator(X_train, variable)
    X_test[variable+'_NA'] = missing_indicator(X_test, variable)
X_train.head()

missing_ind = [col for col in X_train.columns if 'NA' in col]
X_train[missing_ind].mean()

X_train.isnull().mean()

def impute_na(df, variable, value):
    return df[variable].fillna(value)

median = X_train['LotFrontage'].median()
X_train['LotFrontage'] = impute_na(X_train, 'LotFrontage', median)
X_test['LotFrontage'] = impute_na(X_test, 'LotFrontage', median)
median = X_train['MasVnrArea'].median()
X_train['MasVnrArea'] = impute_na(X_train, 'MasVnrArea', median)
X_test['MasVnrArea'] = impute_na(X_test, 'MasVnrArea', median)
mode = X_train['BsmtQual'].mode()[0]
X_train['BsmtQual'] = impute_na(X_train, 'BsmtQual', mode)
X_test['BsmtQual'] = impute_na(X_test, 'BsmtQual', mode)
mode = X_train['FireplaceQu'].mode()[0]
X_train['FireplaceQu'] = impute_na(X_train, 'FireplaceQu', mode)
X_test['FireplaceQu'] = impute_na(X_test, 'FireplaceQu', mode)

X_train.isnull().mean()

As you can see, we have now **the double of features respect to the original dataset. The original dataset had 4 variables, the pre-processed dataset contains 8, plus the target.**
