==> Machine Learning/FEML_Codes/04.11-Frequent-Category-Imputation-Sklearn.ipynb <==
## Frequent category imputation with Scikit-learn ==> SimpleImputer

The **SimpleImputer** class provides **most frequent category imputation** for categorical variables! **Simple to use** if applied to the entire dataframe, it has good quality code, and **fast computation** (numpy), allows for **grid search** over the various imputation techniques and allows for **different missing values encodings** (you can indicate if the missing values are np.nan, or zeroes, etc).
It returns a **numpy array** instead of a pandas dataframe, inconvenient for data analysis, needs to use additional classes to select which features to impute, requires more lines of code!

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
from sklearn.impute import SimpleImputer
from sklearn.compose import ColumnTransformer
from sklearn.pipeline import Pipeline
from sklearn.model_selection import train_test_split

**Use only the following variables for the demo: a mix of categorical and numerical!**

cols_to_use = ['BsmtQual', 'FireplaceQu', 'MSZoning',
               'BsmtUnfSF', 'LotFrontage', 'MasVnrArea',
               'Street', 'Alley', 'SalePrice']

data = pd.read_csv('housingPrices_train.csv', usecols=cols_to_use)
print(data.shape)
data.head()

**Check the null values!..**

data.isnull().mean()

The cateogrical variables Alley, **BsmtQual and FirePlaceQu** contain missing data.

**Separate into training and testing set first let's remove the target from the features!**

cols_to_use.remove('SalePrice')
X_train, X_test, y_train, y_test = train_test_split(data[cols_to_use], # just the features
                                                    data['SalePrice'], # the target
                                                    test_size=0.3, # the percentage of obs in the test set
                                                    random_state=0) # for reproducibility
X_train.shape, X_test.shape

**Check the misssing data again!**

X_train.isnull().mean()

### SimpleImputer on the entire dataset

**Impute the missing values with SimpleImputer! Create an instance! indicate that we want to impute with the most frequent category! The imputer will learn the mode of ALL variables! Categorical or not!**

imputer = SimpleImputer(strategy='most_frequent')
imputer.fit(X_train[cols_to_use])  # fit the imputer

**Look at the learnt frequent values like this:**

imputer.statistics_

**Note** that the transformer learns the most frequent value for both categorical AND numerical variables.

**Investigate the frequent values to corroborate the imputer did a good job!**

X_train[cols_to_use].mode()

**Impute the train and test set! The data will be returned as a numpy array!!!**

X_train = imputer.transform(X_train)
X_test = imputer.transform(X_test)
X_train

**Encode the train set back to a dataframe:**

pd.DataFrame(X_train, columns=cols_to_use).head()

### SimpleImputer: different procedures on different features

X_train, X_test, y_train, y_test = train_test_split(data[cols_to_use],
                                                    data['SalePrice'],
                                                    test_size=0.3,
                                                    random_state=0)
X_train.shape, X_test.shape

**Look at the missing values!**

X_train.isnull().mean()

Now we **impute categorical variables** with the **frequent category**, **numerical variables** with the **mean.**

**Make lists, indicating which features will be imputed with each method! Then put the features list and the transformers together!**

features_numeric = ['BsmtUnfSF', 'LotFrontage', 'MasVnrArea', ]
features_categoric = ['BsmtQual', 'FireplaceQu', 'MSZoning',
                      'Street', 'Alley']
preprocessor = ColumnTransformer(transformers=[
    ('numeric_imputer', SimpleImputer(strategy='mean'), features_numeric),
    ('categoric_imputer', SimpleImputer(strategy='most_frequent'), features_categoric)])

**Fit the preprocessor!**

preprocessor.fit(X_train)

**Explore the transformers like this:**

preprocessor.transformers

**Look at the parameters learnt like this: for the mean imputer!**

preprocessor.named_transformers_['numeric_imputer'].statistics_

**Corroborate the value with that one in the train set!**

X_train[features_numeric].mean()

**For frequent category imputer!**

preprocessor.named_transformers_['categoric_imputer'].statistics_

**Corroborate those values in the train set!**

X_train[features_categoric].mode()

**Impute the data**

X_train = preprocessor.transform(X_train)
X_test = preprocessor.transform(X_test)

X_train.shape

**See how the result of the imputation is a 3 column dataset!**

pd.DataFrame(X_train,
             columns=features_numeric + features_categoric).head()

In this case, we passed all the features available in the dataset to the missing data imputers, so the returned dataset contains all the variables.
