==> Machine Learning/FEML_Codes/04.12-Missing-Category-Imputation-Sklearn.ipynb <==
## Missing Category imputation with Scikit-learn: SimpleImputer

The **SimpleImputer** class of Scikit-learn imputes missing values, by: **mean and median** for numerical variables, **most frequent category** for categorical variables, **arbitrary value** for both categorical and numerical variables. Simple to use! Good quality code! Fast computation! Allows for grid search over the various imputation techniques! Allows for different missing values encodings (you can indicate if the missing values are np.nan, or zeroes, etc)! It returns a numpy array! R instead of a pandas dataframe, inconvenient for data analysis!

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
from sklearn.compose import ColumnTransformer
from sklearn.pipeline import Pipeline
from sklearn.impute import SimpleImputer
from sklearn.model_selection import train_test_split

**Load the dataset with some categorical columns and the target SalePrice!**

cols_to_use = ['BsmtQual', 'FireplaceQu', 'SalePrice']
data = pd.read_csv('housingPrices_train.csv', usecols=cols_to_use)
data.head()

**Inspect the percentage of missing values in each variable!**

data.isnull().mean()

The variables LotFrontage, MasVnrArea and GarageYrBlt contain missing data.

cols_to_use.remove('SalePrice')
X_train, X_test, y_train, y_test = train_test_split(
    data[cols_to_use],  # just the features
    data['SalePrice'],  # the target
    test_size=0.3,  # the percentage of obs in the test set
    random_state=0)  # for reproducibility
X_train.shape, X_test.shape

**Check the misssing data now!**

X_train.isnull().mean()

**Inspect the values of the categorical variable!**

X_train['BsmtQual'].unique()

**Inspect the values of the categorical variable!**

X_train['FireplaceQu'].unique()

**Impute the missing values with SimpleImputer! Create an instance! Indicate that we want to impute by replacing NA with 'Missing'!**

imputer = SimpleImputer(strategy='constant', 
                       fill_value = 'Missing')
imputer.fit(X_train)  # Fit the imputer

**Look at the learnt modes like this:**

imputer.statistics_

**Impute the train and test set! NOTE: the data is a numpy array!!!**

X_train = imputer.transform(X_train)
X_test = imputer.transform(X_test)
X_train

**Encode the train set back to a dataframe:**

X_train = pd.DataFrame(X_train, columns=cols_to_use)
X_train.head()

X_train['BsmtQual'].unique()

X_train.isnull().mean()

**A MASSIVE NOTE OF CAUTION**:

Note that when using SimpleImputer and setting the parameters to: **strategy='constant'** and **fill_value = 'Missing'**. If your dataframe contains variables that are numerical and categorical, NA in both will be replaced by 'Missing" therefore converting your numerical variables into categorical, which is probably not what you are after. Most datasets contain both numerical and categorical variables, so very likely you will have to use a column transformer as shown in previous notebooks and as I also show below again.

**Load the dataset with both numerical and categorical variables!**

cols_to_use = ['BsmtQual', 'FireplaceQu', 'LotFrontage',
               'MasVnrArea', 'GarageYrBlt', 'SalePrice']
data = pd.read_csv('housingPrices_train.csv', usecols=cols_to_use)
data.head()

cols_to_use.remove('SalePrice')
X_train, X_test, y_train, y_test = train_test_split(data[cols_to_use],
                                                    data['SalePrice'],
                                                    test_size=0.3,
                                                    random_state=0)
X_train.shape, X_test.shape

**Look at the missing values!**

X_train.isnull().mean()

For this demo, I will impute the numerical variables by the mean, and the categorical variables by the most frequent category.

**Make lists, indicating which features will be imputed with each method!**

features_numeric = ['LotFrontage', 'MasVnrArea', 'GarageYrBlt']
features_categoric = ['BsmtQual', 'FireplaceQu']
preprocessor = ColumnTransformer(transformers=[ # put the features list and the transformers together
    ('imputer_numeric', SimpleImputer(strategy='mean'), features_numeric),
    ('imputer_categoric', SimpleImputer(strategy='constant', fill_value='Missing'), features_categoric)])

**Fit the preprocessor!**

preprocessor.fit(X_train)

**Explore the transformers like this:**

preprocessor.transformers

**Look at the parameters learnt like this: for the numerical imputer!**

preprocessor.named_transformers_['imputer_numeric'].statistics_

**For the categorical imputer**

preprocessor.named_transformers_['imputer_categoric'].statistics_

**Impute the data! Data is a numpy array!**

X_train = preprocessor.transform(X_train)
X_test = preprocessor.transform(X_test)

**Convert the result into a dataframe!**

pd.DataFrame(X_train,
             columns=features_numeric+features_categoric).head()

**Convert the result into a dataframe! Explore the missing values, there should be none!**

X_train = pd.DataFrame(X_train,
             columns=features_numeric+features_categoric)
X_train.isnull().mean()
