==> Machine Learning/FEML_Codes/04.13-MissingIndicator-Sklearn.ipynb <==
## Adding a Missing Indicator variable with Scikit-learn ==> MissingIndicator

Scikit-learn provides the **MissingIndicator** class to add a binary variable that flags NA. The MissingIndicator has the option of adding a Missing indicator binary variable to all the variables in the dataset, or only those that show NA in the train set. **Attention!** The transformer only returns the binary variables, which need to be added to the original train data.

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
from sklearn.impute import SimpleImputer, MissingIndicator
from sklearn.model_selection import train_test_split

**A mix of categorical and numerical variables!**

cols_to_use = ['BsmtQual', 'FireplaceQu', 'MSZoning',
               'BsmtUnfSF', 'LotFrontage', 'MasVnrArea',
               'Street', 'Alley', 'SalePrice']

**Load the House Prices dataset!**

data = pd.read_csv('housingPrices_train.csv', usecols=cols_to_use)
print(data.shape)
data.head()

**Check the null values**

data.isnull().mean()

cols_to_use.remove('SalePrice')
X_train, X_test, y_train, y_test = train_test_split(data[cols_to_use], # just the features
                                                    data['SalePrice'], # the target
                                                    test_size=0.3, # the percentage of obs in the test set
                                                    random_state=0) # for reproducibility
X_train.shape, X_test.shape

**Check the misssing data again!**

X_train.isnull().mean()

## Add a Missing Indicator

indicator = MissingIndicator(error_on_new=True, features='missing-only')
indicator.fit(X_train)  

**See the features with na: the result shows the index!**

indicator.features_

**Find the feature names by passing the index to the list of columns**

X_train.columns[indicator.features_]

**The indicator returns only the additional indicators when we transform the dataset!**

tmp = indicator.transform(X_train)
tmp

**Join it manually to the original X_train! Create a column name for each of the new MissingIndicators!**

indicator_cols = [c+'_NA' for c in X_train.columns[indicator.features_]]
X_train = pd.concat([  # Now concatenate
    X_train.reset_index(),
    pd.DataFrame(tmp, columns = indicator_cols)],
    axis=1)
X_train.head()

**Now the same for the test set**

tmp = indicator.transform(X_test)
X_test = pd.concat([
    X_test.reset_index(),
    pd.DataFrame(tmp, columns = indicator_cols)],
    axis=1)
X_test.head()

### SimpleImputer on the entire dataset

**Impute the missing values with SimpleImputer! Create an instance! Indicate that we want to impute with the 
most frequent category!**

imputer = SimpleImputer(strategy='most_frequent')
imputer.fit(X_train)  # Fit the imputer to the train set

**Look at the learnt frequent values like this:**

imputer.statistics_

**Note** that the transformer learns the most frequent value for both categorical AND numerical variables.

**Impute the train and test set! Data is a numpy array!!!**

X_train = imputer.transform(X_train)
X_test = imputer.transform(X_test)
X_train
