==> Machine Learning/FEML_Codes/04.14-Automatic-Imputation-Method-Detection-Sklearn.ipynb <==
## Automatic selection of best imputation technique with Sklearn

Now we will do a **grid search** over the imputation methods available in Scikit-learn to determine **which imputation technique works best** for this dataset and the machine learning model of choice. We will also train a **very simple machine learning model** as part of a small pipeline.

import pandas as pd
import numpy as np
from sklearn.compose import ColumnTransformer
from sklearn.pipeline import Pipeline
from sklearn.impute import SimpleImputer
from sklearn.preprocessing import StandardScaler, OneHotEncoder
from sklearn.linear_model import Lasso
from sklearn.model_selection import train_test_split, GridSearchCV
np.random.seed(0)

**Load dataset with all the variables!**

data = pd.read_csv('housingPrices_train.csv',)
data.head()

**Find categorical and numerical variables! Categorical variables' type is 'Object' in the dataset!**

features_categorical = [c for c in data.columns if data[c].dtypes=='O']
features_numerical = [c for c in data.columns if data[c].dtypes!='O' and c !='SalePrice']

**Inspect the categorical variables!**

data[features_categorical].head()

**Inspect the numerical variables!**

data[features_numerical].head()

X_train, X_test, y_train, y_test = train_test_split(
    data.drop('SalePrice', axis=1),  # just the features
    data['SalePrice'],  # the target
    test_size=0.3,  # the percentage of obs in the test set
    random_state=0)  # for reproducibility
X_train.shape, X_test.shape

**Create the preprocessing pipelines for both numerical and categorical data!**

numeric_transformer = Pipeline(steps=[
    ('imputer', SimpleImputer(strategy='median')),
    ('scaler', StandardScaler())])
categorical_transformer = Pipeline(steps=[
    ('imputer', SimpleImputer(strategy='constant', fill_value='missing')),
    ('onehot', OneHotEncoder(handle_unknown='ignore'))])
preprocessor = ColumnTransformer(
    transformers=[
        ('numerical', numeric_transformer, features_numerical),
        ('categorical', categorical_transformer, features_categorical)])

We passed any argument to the transformers those will be changed during the gridsearch below!

**Append classifier to preprocessing pipeline. Now we have a full prediction pipeline!**

clf = Pipeline(steps=[('preprocessor', preprocessor),
                      ('regressor', Lasso(max_iter=2000))])

**Create the grid with all the parameters that we would like to test!**

param_grid = {
    'preprocessor__numerical__imputer__strategy': ['mean', 'median'], # mean and median for numerical processor!
    'preprocessor__categorical__imputer__strategy': ['most_frequent', 'constant'], # These are for categorical processor!
    'regressor__alpha': [10, 100, 200],}  # classifier__alpha': [0.1, 1.0, 0.5] !
grid_search = GridSearchCV(clf, param_grid, cv=5, n_jobs=-1, scoring='r2')the above line of code indicates that I want to test those 3 values for the alpha parameter of Lasso. Note that Lasso is the 'classifier' step of our last pipeline

**cv=3** is the cross-validation, **no_jobs** =-1 indicates to use all available cpus, **scoring='r2'** indicates to evaluate using the r squared! 

classifier__alpha': [0.1, 1.0, 0.5]  --> This line of code indicates that I want to test those 3 values for the alpha parameter of Lasso. Note that Lasso is the 'classifier' step of our last pipeline!

**Train over all the possible combinations of the parameters above!**

grid_search.fit(X_train, y_train)
print(("best linear regression from grid search: %.3f"
       % grid_search.score(X_train, y_train)))  #prints the best score!

**Print the best estimator parameters like this!**

grid_search.best_estimator_

**Find the best fit parameters like this!**

grid_search.best_params_

**See all the combinations evaluated during the gridsearch!**

grid_search.cv_results_['params']

**Here the scores for each of one of the above combination!s**

grid_search.cv_results_['mean_test_score']

**Finally let's check the performance over the test set!**

print(("best linear regression from grid search: %.3f"
       % grid_search.score(X_test, y_test)))

This model overfits to the train set, look at the r2 of 0.93 obtained for the train set vs 0.738 for the test set.

We will try to reduce this over-fitting as we progress in the course.
