==> Machine Learning/FEML_Codes/04.16-Mean-Median-Imputation-Feature-Engine.ipynb <==
## Mean / Median Imputation ==> Feature-Engine

**Feature-Engine**, an **open source python package**, includes all the **feature engineering techniques**, **works like Scikit-learn** (easy to learn), allows you to implement specific engineering steps to specific feature subsets, and can be **integrated with the Scikit-learn pipeline** allowing for smooth model building! Feature-Engine allows you to design and store a feature engineering pipeline with bespoke procedures for different variable groups.

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
from sklearn.model_selection import train_test_split
from sklearn.pipeline import Pipeline
from feature_engine.imputation import MeanMedianImputer

cols_to_use = ['BsmtQual', 'FireplaceQu', 'LotFrontage', 
               'MasVnrArea', 'GarageYrBlt', 'SalePrice']
data = pd.read_csv('housingPrices_train.csv', usecols=cols_to_use)
data.head()

data.isnull().mean()

All the predictor variables contain missing data.

cols_to_use.remove('SalePrice')
X_train, X_test, y_train, y_test = train_test_split(data[cols_to_use],
                                                    data['SalePrice'],
                                                    test_size=0.3,
                                                    random_state=0)
X_train.shape, X_test.shape

### Feature-Engine captures the numerical variables automatically

**Imputer from feature-engine! Specify the imputation strategy, median in this case!**

imputer = MeanMedianImputer(imputation_method='median')

**Fit the imputer!**

imputer.fit(X_train)

**Imputer found the numerical variables to impute with the mean!**

imputer.variables_

**Check the mean assigned to each variable!**

imputer.imputer_dict_

**feature-engine returns a dataframe!**

tmp = imputer.transform(X_train)
tmp.head()

**Check that the numerical variables don't contain NA any more!**

## Feature-Engine allows you to specify variable groups easily

**Mean imputation! Do it over 2 of the 3 numerical variables!**

imputer = MeanMedianImputer(imputation_method='mean',
                            variables=['LotFrontage', 'MasVnrArea'])
imputer.fit(X_train)

**Now the imputer uses only the variables we indicated!**

imputer.variables_

**Corroborate that the dictionary above contains the mean values of th variables!**

X_train[imputer.variables_].mean()

**feature-engine returns a dataframe!**

tmp = imputer.transform(X_train)
tmp[imputer.variables_].isnull().mean()  # let's check null values are gone

## Feature-Engine can be used with the Scikit-learn pipeline

pipe = Pipeline([
    ('median_imputer', MeanMedianImputer(imputation_method='median',
                                         variables=['LotFrontage', 'GarageYrBlt'])),
    ('mean_imputer', MeanMedianImputer(imputation_method='mean',
                                       variables=['MasVnrArea'])),])

pipe.fit(X_train)

pipe.named_steps['median_imputer'].imputer_dict_

pipe.named_steps['mean_imputer'].imputer_dict_

*Transform the data with the pipeline! Check null values are gone!**

tmp = pipe.transform(X_train)
tmp.isnull().mean()
