==> Machine Learning/FEML_Codes/04.20-Missing-Category-Imputation-Feature-Engine.ipynb <==
## Missing Category Imputation ==> Feature-Engine

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
from sklearn.model_selection import train_test_split
from sklearn.pipeline import Pipeline
from feature_engine.imputation import CategoricalImputer

cols_to_use = ['BsmtQual', 'FireplaceQu', 'LotFrontage', 'MasVnrArea', 'GarageYrBlt', 'SalePrice']
data = pd.read_csv('housingPrices_train.csv', usecols=cols_to_use)
data.head()

data.isnull().mean()

cols_to_use.remove('SalePrice')
X_train, X_test, y_train, y_test = train_test_split(data[cols_to_use],
                                                    data['SalePrice'],
                                                    test_size=0.3,
                                                    random_state=0)
X_train.shape, X_test.shape

### Feature-Engine captures the categorical variables automatically

**Call the imputer from feature- engine! No need to specify anything!**

imputer = CategoricalImputer()

**Fit the imputer!**

imputer.fit(X_train)

**The imputer finds the categorical variables to impute with the frequent category!**

imputer.variables_

**This imputer will replace missing data in categorical variables by 'Missing'**

**feature-engine returns a dataframe!**

tmp = imputer.transform(X_train)
tmp.head()

**Check that the numerical variables don't contain NA any more!**

tmp[imputer.variables_].isnull().mean()

## Feature-engine allows you to specify variable groups easily

**Do imputation over 1 of the 2 categorical variables!**

imputer = CategoricalImputer(variables=['BsmtQual'])
imputer.fit(X_train)

imputer.variables_

**Replace missing data by 'missing'**.

tmp = imputer.transform(X_train)
tmp.head()

**Check it not to contain NA anymore!**

tmp[imputer.variables_].isnull().mean()

## Feature-engine can be used with the Scikit-learn pipeline

**M**

**Check the percentage of NA in each categorical variable!**

X_train.isnull().mean()

- BsmtQual: 0.023 ==> frequent category imputation
- FirePlaceQu: 0.46 ==> missing category imputation

pipe = Pipeline([
    ('imputer_mode', CategoricalImputer(imputation_method='frequent', variables=['BsmtQual'])),
    ('imputer_missing', CategoricalImputer(variables=['FireplaceQu'])),])

pipe.fit(X_train)

pipe.named_steps['imputer_mode'].variables

pipe.named_steps['imputer_missing'].variables

**Transform the data with the pipeline! Check null values are gone!**

tmp = pipe.transform(X_train)
tmp.isnull().mean()
