==> Machine Learning/FEML_Codes/04.22-Missing-Indicator-Feature-Engine.ipynb <==
## Missing Indicator ==> Feature-Engine

**Feature-Engine**, an **open source python package**, includes all the **feature engineering techniques**,  works **like Scikit-learn** (easy to learn), allows you to implement specific engineering steps to specific feature subsets, and can be **integrated with the Scikit-learn pipeline** allowing for smooth model building! **Feature-Engine allows you to design and store a feature engineering pipeline with bespoke procedures for different variable groups.**

**# !pip install feature-engine**

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
from sklearn.model_selection import train_test_split
from sklearn.pipeline import Pipeline
from feature_engine.imputation import (
    AddMissingIndicator,
    MeanMedianImputer,
    CategoricalImputer)

cols_to_use = ['BsmtQual', 'FireplaceQu', 'LotFrontage', 'MasVnrArea', 'GarageYrBlt', 'SalePrice']
data = pd.read_csv('housingPrices_train.csv', usecols=cols_to_use)
data.head()

data.isnull().mean()

cols_to_use.remove('SalePrice')
X_train, X_test, y_train, y_test = train_test_split(data[cols_to_use],
                                                    data['SalePrice'],
                                                    test_size=0.3,
                                                    random_state=0)
X_train.shape, X_test.shape

### Feature-Engine's missing indicator selects all variables by default

**Call the imputer from feature-engine! the argument  allows us how to determine if we wantto add missing indicators!..**

imputer = AddMissingIndicator(missing_only=True)

**Fit the imputer!**

imputer.fit(X_train)

**The attribute variables shows the variables entered by the user in this None!!**

imputer.variables

**This attribute stores the variables, numerical and categorical, if they showed missing data in the train set!**

imputer.variables_

**feature-engine returns a dataframe with the additional features! No need to contatenate!!**

tmp = imputer.transform(X_train)
tmp.head()

**Check NA!**

tmp.isnull().mean()

## Feature-engine allows you to specify variable groups easily

**Do it imputation but this time! Select a few variables!**

imputer = AddMissingIndicator(variables=['BsmtQual', 'FireplaceQu', 'LotFrontage'])
imputer.fit(X_train)

**Now the imputer uses only the variables we indicated!**

imputer.variables

**Missing indicators** will be added for the following variables in case that these are **different from the ones passed by the user!**

imputer.variables_

**Feature-engine returns a dataframe with the additional features! No need to contatenate!!**

tmp = imputer.transform(X_train)
tmp.head()

## Feature-engine can be used with the Scikit-learn pipeline

**feature-engine returns a dataframe with the additional features! No need to contatenate!!**

**M**

X_train.isnull().mean()

X_train.head()

**Check the percentage of NA in each variable!**

These are the steps we will concatenate: **Add Missing Indicator to all variables! Median Imputation to numerical variables! Frequent category imputation to categorical variables!**

pipe = Pipeline([
    ('missing_ind', AddMissingIndicator()),
    ('imputer_mode', CategoricalImputer(
        imputation_method='frequent', variables=['FireplaceQu', 'BsmtQual'])),
    ('imputer_median', MeanMedianImputer(imputation_method='median',
                                         variables=['LotFrontage', 'MasVnrArea', 'GarageYrBlt'])),])

**Fit the pipe!**

pipe.fit(X_train)

**Inspect the separate steps!...**

pipe.named_steps['missing_ind'].variables_

pipe.named_steps['imputer_mode'].imputer_dict_

pipe.named_steps['imputer_median'].imputer_dict_

**Transform the data with the pipeline!** It will **add the missing indicators**, **fill na** in the original variables, **make the dataset ready to use for ML !!**

tmp = pipe.transform(X_train)
tmp.isnull().mean()  # let's check null values are gone

tmp.shape
