==> Machine Learning/FEML_Codes/05.01-KNN-imputation.ipynb <==
## KNN imputation

**The missing values** are estimated as the **average value from the closest K neighbours**. [KNNImputer from sklearn](https://scikit-learn.org/stable/modules/generated/sklearn.impute.KNNImputer.html#sklearn.impute.KNNImputer): **Same K will be used** to impute all variables! Can't really optimise K to better predict the missing values! Could optimise K to better predict the target! **Note** If what we want is to predict, **as accurately as possible** the values of the missing data, then, we would not use the KNN imputer, we would build individual KNN algorithms to predict 1 variable from the remaining ones. This is a common regression problem.

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
from sklearn.model_selection import train_test_split
from sklearn.impute import KNNImputer  # multivariate imputation

## Load data

**List with numerical varables!**

cols_to_use = [
    'MSSubClass', 'LotFrontage', 'LotArea', 'OverallQual', 'OverallCond', 
    'YearBuilt', 'YearRemodAdd', 'MasVnrArea', 'BsmtFinSF1', 'BsmtFinSF2', 
    'BsmtUnfSF', 'TotalBsmtSF', '1stFlrSF', '2ndFlrSF', 'LowQualFinSF', 
    'GrLivArea', 'BsmtFullBath', 'BsmtHalfBath', 'FullBath', 'HalfBath',
    'BedroomAbvGr', 'KitchenAbvGr', 'TotRmsAbvGrd', 'Fireplaces', 'GarageYrBlt', 
    'GarageCars', 'GarageArea', 'WoodDeckSF',  'OpenPorchSF', 'EnclosedPorch', 
    '3SsnPorch', 'ScreenPorch', 'PoolArea', 'MiscVal', 'MoSold', 'YrSold', 'SalePrice']

**Load the dataset with a selected variables!**

data = pd.read_csv('HousingPrices_train.csv', usecols=cols_to_use)
data.head()

for var in data.columns:  # find variables with missing data
    if data[var].isnull().sum() > 1:
        print(var, data[var].isnull().sum())

cols_to_use.remove('SalePrice')
X_train, X_test, y_train, y_test = train_test_split(
    data[cols_to_use],
    data['SalePrice'],
    test_size=0.3,
    random_state=0)
X_train.shape, X_test.shape

**Reset index, so we can compare values later!**

X_train.reset_index(inplace=True, drop=True)
X_test.reset_index(inplace=True, drop=True)

## KNN imputation

imputer = KNNImputer(
    n_neighbors=5, # the number of neighbours K
    weights='distance', # the weighting factor
    metric='nan_euclidean', # the metric to find the neighbours
    add_indicator=False,)  # whether to add a missing indicator

imputer.fit(X_train)

**Sklearn returns a Numpy array! Make a dataframe!**

train_t = imputer.transform(X_train)
test_t = imputer.transform(X_test)
train_t = pd.DataFrame(train_t, columns=X_train.columns)
test_t = pd.DataFrame(test_t, columns=X_test.columns)
train_t.head()

**Variables without NA after the imputation!**

train_t[['LotFrontage', 'MasVnrArea', 'GarageYrBlt']].isnull().sum()

**The obseravtions with NA in the original train set!**

X_train[X_train['MasVnrArea'].isnull()]['MasVnrArea']

**The replacement values in the transformed dataset!**

train_t[X_train['MasVnrArea'].isnull()]['MasVnrArea']

**The mean value of the variable (i.e., for mean imputation)!**

X_train['MasVnrArea'].mean()

In some cases, the imputation values are **very different from the mean value** we would have used in MeanMedianImputation.

## Imputing a slice of the dataframe

We can use **Feature-engine** to apply the KNNImputer to **a slice of the dataframe.**

from feature_engine.wrappers import SklearnTransformerWrapper

data = pd.read_csv('HousingPrices_train.csv')
X_train, X_test, y_train, y_test = train_test_split(
    data.drop('SalePrice', axis=1),
    data['SalePrice'],
    test_size=0.3,
    random_state=0)
X_train.shape, X_test.shape

**Start the KNNimputer inside the SKlearnTransformerWrapper!**

imputer = SklearnTransformerWrapper(
    transformer = KNNImputer(weights='distance'),
    variables = cols_to_use,)

**Fit the wrapper + KNNImputer! Transform the data! Feature-engine returns a dataframe!**

imputer.fit(X_train)
train_t = imputer.transform(X_train)
test_t = imputer.transform(X_test)
train_t.head()

**No NA after the imputation!**

train_t['MasVnrArea'].isnull().sum()

**Same imputation values as previously!**

train_t[X_train['MasVnrArea'].isnull()]['MasVnrArea']

## Automatically find best imputation parameters

We can **optimise the parameters of the KNN imputation** to better predict our outcome.

**Import extra classes for modelling!**

from sklearn.preprocessing import StandardScaler
from sklearn.linear_model import Lasso
from sklearn.pipeline import Pipeline
from sklearn.model_selection import GridSearchCV

**Separate intro train and test set!..**

X_train, X_test, y_train, y_test = train_test_split(
    data[cols_to_use],  # just the features
    data['SalePrice'],  # the target
    test_size=0.3,  # the percentage of obs in the test set
    random_state=0)  # for reproducibility
X_train.shape, X_test.shape

pipe = Pipeline(steps=[
    ('imputer', KNNImputer(
        n_neighbors=5,
        weights='distance',
        add_indicator=False)),
    ('scaler', StandardScaler()),
    ('regressor', Lasso(max_iter=2000)),])

**Create the grid with all the parameters that we would like to test!**

param_grid = {
    'imputer__n_neighbors': [3,5,10],
    'imputer__weights': ['uniform', 'distance'],
    'imputer__add_indicator': [True, False],
    'regressor__alpha': [10, 100, 200],}
grid_search = GridSearchCV(pipe, param_grid, cv=5, n_jobs=-1, scoring='r2')

**cv=3** is the cross-validation, **no_jobs =-1** indicates to use all available cpus, **scoring='r2'** indicates to evaluate using the r squared. **For more details in the grid parameters** visit: #https://scikit-learn.org/stable/modules/generated/sklearn.model_selection.GridSearchCV.html

**Train over all the possible combinations of the parameters above!**

grid_search.fit(X_train, y_train)
print(("best linear regression from grid search: %.3f"
       % grid_search.score(X_train, y_train)))  # best score!

**Check the performance over the test set print!**

print(("best linear regression from grid search: %.3f"
       % grid_search.score(X_test, y_test)))

**Find the best parameters!**

grid_search.best_params_

## Compare with univariate imputation

from sklearn.impute import SimpleImputer

**Separate intro train and test set!**

X_train, X_test, y_train, y_test = train_test_split(
    data[cols_to_use],  # just the features
    data['SalePrice'],  # the target
    test_size=0.3,  # the percentage of obs in the test set
    random_state=0)  # for reproducibility
X_train.shape, X_test.shape

**Train over all the possible combinations of the given parameters!**

pipe = Pipeline(steps=[
    ('imputer', SimpleImputer(strategy='mean', fill_value=-1)),
    ('scaler', StandardScaler()),
    ('regressor', Lasso(max_iter=2000)),])
param_grid = {
    'imputer__strategy': ['mean', 'median', 'constant'],
    'imputer__add_indicator': [True, False],
    'regressor__alpha': [10, 100, 200],}
grid_search = GridSearchCV(pipe, param_grid, cv=5, n_jobs=-1, scoring='r2')
grid_search.fit(X_train, y_train)
print(("best linear regression from grid search: %.3f"
       % grid_search.score(X_train, y_train)))

**Check the performance over the test set!**

print(("best linear regression from grid search: %.3f"
       % grid_search.score(X_test, y_test)))

**Find the best fit parameters like this!**

grid_search.best_params_

We see that imputing the values with an **arbitrary value of -1**, returns approximately the **same performance as doing KNN imputation**, so we might not want to add the additional complexity of training models to impute NA, too, then go ahead and predict the real target we are interested in.
