==> Machine Learning/FEML_Codes/06.01-One-hot-encoding.ipynb <==
## One Hot Encoding

**One hot encoding**, consists in **encoding each categorical variable with different boolean variables** (also called **dummy variables**) which take values 0 or 1, indicating **if a category is present in an observation**, like "Gender", with labels 'female' and 'male', assigned as 0 and 1, or 'color' with color labels as numbers!

**When one hot encoding categorical variables, we create k - 1 binary variables** For example, for **blue, red and green** colors, only 2 binary variable is sufficient, **if green, red and blue should be 0**. no need the green variable! If k=2 (male/female), only 1 variable (k-1=1) is sufficient!

But, to build **tree based algorithms** or do feature selection by **recursive algorithms**, or we want to determine **the importance of each single category**, we should do **one hot encoding into k dummy variables**.

It is **Straightforward to implement!** Makes **no assumption** about the distribution or categories! **Keeps all the information** of the categorical variable! Suitable for linear models! But it **expands the feature space**! Does not add extra information while encoding! Many dummy variables may be identical, introducing redundant information!

import pandas as pd
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import OneHotEncoder  # one hot encoding with sklearn
from feature_engine.encoding import OneHotEncoder as fe_OneHotEncoder # with feature-engine

**Load titanic dataset!**

data = pd.read_csv('titanic.csv',
                   usecols=['sex', 'embarked', 'cabin', 'survived'])
data.head()

**Capture only the first letter of the cabin for this demonstration!**

data['cabin'] = data['cabin'].str[0]
data.head()

**Separate into training and testing set!**

X_train, X_test, y_train, y_test = train_test_split(
    data[['sex', 'embarked', 'cabin']],  # predictors
    data['survived'],  # target
    test_size=0.3,  # percentage of obs in test set
    random_state=0)  # seed to ensure reproducibility
X_train.shape, X_test.shape

### Let's explore the cardinality

**Sex has 2 labels!**

X_train['sex'].unique()

**Embarked has 3 labels and missing data!**

X_train['embarked'].unique()

**Cabin has 9 labels and missing data!**

X_train['cabin'].unique()

## One hot encoding with pandas

It is **quick**, returns pandas **dataframe**, returns **feature names for the dummy variables**. But it does **not preserve information from train data** to propagate to test data! Pandas.get_dummies(), will create as many binary variables as categories in the variable!

### into k  dummy variables

**Create dummy variables with the build in pandas method get_dummies!**

tmp = pd.get_dummies(X_train['sex'])
tmp.head()

**For better visualisation let's put the dummies next to the original variable!**

pd.concat([X_train['sex'],
           pd.get_dummies(X_train['sex'])], axis=1).head()

**Repeat for embarked!**

tmp = pd.get_dummies(X_train['embarked'])
tmp.head()

**For better visualisation!**

pd.concat([X_train['embarked'],
           pd.get_dummies(X_train['embarked'])], axis=1).head()

**Now for cabin!**

tmp = pd.get_dummies(X_train['cabin'])
tmp.head()

**Now for all variables together: train set!**

tmp = pd.get_dummies(X_train)
print(tmp.shape)
tmp.head()

**Now for all variables together: test set!!**

tmp = pd.get_dummies(X_test)
print(tmp.shape)
tmp.head()

Notice the positives of pandas get_dummies: dataframe returned with **feature names!** But the train set contains **13 dummy features**, whereas the test set contains **12 features**. This occurred because there was no category T in cabin in the test set. Predictors require train and test sets to be of the same shape!

### into k -1 

**obtain k-1 labels:** we need to indicate **get_dummies to drop the first binary variable!**

tmp = pd.get_dummies(X_train['sex'], drop_first=True)
tmp.head()

tmp = pd.get_dummies(X_train['embarked'], drop_first=True)
tmp.head()

For embarked, if an observation shows **0 for Q and S**, then **its value must be C**, the remaining category!

tmp = pd.get_dummies(X_train, drop_first=True)
print(tmp.shape)  # altogether: train set
tmp.head()

tmp = pd.get_dummies(X_test, drop_first=True)
print(tmp.shape)   # altogether: train set
tmp.head()

### Bonus: get_dummies() can handle missing values

pd.get_dummies(X_train['embarked'], drop_first=True, dummy_na=True).head()

## One hot encoding with Scikit-learn

It is **quick**, creates the **same number of features in train and test set**. but it returns a **numpy array** instead of a pandas dataframe, it does **not return the variable names**, therefore inconvenient for variable exploration!

**Create and train the encoder!**

encoder = OneHotEncoder(categories='auto',
                       drop='first', # to return k-1, use drop=false to return k dummies
                       sparse=False,
                       handle_unknown='error') # helps deal with rare labels
encoder.fit(X_train.fillna('Missing'))

**Observe the learned categories!**

encoder.categories_

**Transform the train set!..**

tmp = encoder.transform(X_train.fillna('Missing'))
pd.DataFrame(tmp).head()

**NEW: in latest release of Scikit-learn!Now retrieve the feature names as follows:**

encoder.get_feature_names()

**Transfom the test set! Then reconstitute it back to a dataframe! Add the feature names derived by OHE!**

tmp = encoder.transform(X_test.fillna('Missing'))
tmp = pd.DataFrame(tmp)
tmp.columns = encoder.get_feature_names()
tmp.head()

We can see that **train and test contain the same number** of features!


## One hot encoding with Feature-Engine

It is **quick**, returns **dataframe**, returns **feature names**, allows to **select features to encode**! **Limitations** Not sure yet!

ohe_enc = fe_OneHotEncoder(
    top_categories=None,
    variables=['sex', 'embarked'],  # we can select which variables to encode
    drop_last=True)  # to return k-1, false to return k
ohe_enc.fit(X_train.fillna('Missing'))

tmp = ohe_enc.transform(X_train.fillna('Missing'))
tmp.head()

Note how feature-engine returns the dummy variables with their names, and drops the original variable, leaving the dataset ready for further exploration or building machine learning models.

tmp = ohe_enc.transform(X_test.fillna('Missing'))
tmp.head()

**Feature-Engine's one hot encoder also selects all categorical variables automatically!**

ohe_enc = fe_OneHotEncoder(
    top_categories=None,
    drop_last=True)  # to return k-1, false to return k
ohe_enc.fit(X_train.fillna('Missing'))

ohe_enc.variables_

tmp = ohe_enc.transform(X_train.fillna('Missing'))
tmp.head()

tmp = ohe_enc.transform(X_test.fillna('Missing'))
tmp.head()

Note how this encoder returns a variable cabin_T for the test set as well, even though this category is not present in the test set. This allows the integration with Scikit-learn pipeline and scoring of test set by the built algorithm..

In fact, we can check that the sum of cabin_t is 0:

tmp['cabin_T'].sum()
