==> Machine Learning/FEML_Codes/06.02-One-hot-encoding-frequent_categories.ipynb <==
## One Hot Encoding of Frequent Categories

Rare labels may appear only in the train set, therefore cause over-fitting, or only in the test set, and then our models wouldn't know how to score those observations. **In order to avoid these complications, we can create dummy variables only for the most frequent categories**

**OHE of top categories** Straightforward to implement! Does not require hrs of variable exploration! Does not expand massively the feature space! Suitable for linear models! But it does **not add any information that may make the variable more predictive**,  and does **not keep the information of the ignored labels**.

### Note

The number of top variables is set arbitrarily.** In the KDD competition the authors selected **10**, but it could have been **15 or 5** as well. 

import numpy as np
import pandas as pd
from sklearn.model_selection import train_test_split
from feature_engine.encoding import OneHotEncoder

data = pd.read_csv('HousingPrices_train.csv',
                   usecols=['Neighborhood', 'Exterior1st', 'Exterior2nd', 'SalePrice'])
data.head()

**Check how many labels each variable has!**

for col in data.columns:
    print(col, ': ', len(data[col].unique()), ' labels')

**Explore the unique categories!**

data['Neighborhood'].unique()

data['Exterior1st'].unique()

data['Exterior2nd'].unique()

### Encoding important

It is important to select the top or most frequent categories based of the train data. Then, we will use those top categories to encode the variables in the test data as well

**Separate into training and testing set!**

X_train, X_test, y_train, y_test = train_test_split(
    data[['Neighborhood', 'Exterior1st', 'Exterior2nd']],  # predictors
    data['SalePrice'],  # target
    test_size=0.3,  # percentage of obs in test set
    random_state=0)  # seed to ensure reproducibility
X_train.shape, X_test.shape

**First examine how OHE expands the feature space!**

pd.get_dummies(X_train, drop_first=True).shape

**From the initial 3 categorical variables, we end up with 53 variables.**
In real-life datasets, categorical variables can be highly cardinal, and with OHE we can end up with datasets with **thousands of columns.**


OHE with pandas and NumPy is quick, returns pandas dataframe, returns feature names for the dummy variables! But it does **not preserve information from train data** to propagate to test data!

**Find the top 10 most frequent categories for the variable 'Neighborhood'!**

X_train['Neighborhood'].value_counts().sort_values(ascending=False).head(10)

**Make a list with the most frequent categories of the variable!**

top_10 = [x for x in X_train['Neighborhood'].value_counts().sort_values(
        ascending=False).head(10).index]
top_10

**Make the 10 binary variables" and Visualise the result**

for label in top_10:
    X_train['Neighborhood' + '_' + label] = np.where(
        X_train['Neighborhood'] == label, 1, 0)
    X_test['Neighborhood' + '_' + label] = np.where(
        X_test['Neighborhood'] == label, 1, 0)
X_train[['Neighborhood'] + ['Neighborhood'+'_'+c for c in top_10]].head(10)

**We can turn the previous commands into 2 functions!**

def calculate_top_categories(df, variable, how_many=10):
    return [x for x in df[variable].value_counts().sort_values(
            ascending=False).head(how_many).index]

def one_hot_encode(train, test, variable, top_x_labels):
    for label in top_x_labels:
        train[variable + '_' + label] = np.where(
            train[variable] == label, 1, 0)
        test[variable + '_' + label] = np.where(
            test[variable] == label,1, 0)

**Now we run a loop over the remaining categorical variables!**

for variable in ['Exterior1st', 'Exterior2nd']:
    top_categories = calculate_top_categories(X_train, variable, how_many=10)
    one_hot_encode(X_train, X_test, variable, top_categories)

**Let's see the result!**

X_train.head()

**Check the result!**

Note how we now have **30 additional dummy variables instead of the 53** that we would have had if we had created dummies for all categories.

## One hot encoding of top categories with Feature-Engine

It is **quick** and **creates the same number of features** in train and test set! No known disadvantage!

**Separate into training and testing set!**

X_train, X_test, y_train, y_test = train_test_split(
    data[['Neighborhood', 'Exterior1st', 'Exterior2nd']],  # predictors
    data['SalePrice'],  # target
    test_size=0.3,  # percentage of obs in test set
    random_state=0)  # seed to ensure reproducibility
X_train.shape, X_test.shape

ohe_enc = OneHotEncoder(
    top_categories=10,  # you can change this value to select more or less variables
    # we can select which variables to encode
    variables=['Neighborhood', 'Exterior1st', 'Exterior2nd'],
    drop_last=False)

ohe_enc.fit(X_train)

**In the encoder dict we can observe each of the top categories selected for each of the variables!**

ohe_enc.encoder_dict_

**The list of variables that the encoder will transform!**

ohe_enc.variables_

X_train = ohe_enc.transform(X_train)
X_test = ohe_enc.transform(X_test)
X_train.head()  # let's explore the result

**Note**

If the argument variables is left to None, then the encoder will automatically identify **all categorical variables**. It is very good! The encoder will **not encode numerical variables**. So if some of your numerical variables are in fact categories, you will need to re-cast them as object before using the encoder.
