==> Machine Learning/FEML_Codes/06.03-Integer-Encoding.ipynb <==
## Integer Encoding

Integer encoding consist in **replacing the categories by digits** from 1 to n (or 0 to n-1, depending the implementation), where n is the number of distinct categories of the variable. The numbers are **assigned arbitrarily.** This encoding method allows for quick benchmarking of machine learning models. **Advantages:** Straightforward to implement! Does not expand the feature space! **Limitations:** Does not capture any information about the categories labels! Not suitable for linear models!

import numpy as np
import pandas as pd
from sklearn.model_selection import train_test_split
from sklearn.preprocessing import LabelEncoder
from feature_engine.encoding import OrdinalEncoder

data = pd.read_excel('HousingPrices.xls', 
    usecols=['Neighborhood', 'Exterior1st', 'Exterior2nd', 'SalePrice'])
data.head()

**Check how many labels each variable has!**

for col in data.columns:
    print(col, ': ', len(data[col].unique()), ' labels')

**Explore the unique categories!**

data['Neighborhood'].unique()

data['Exterior1st'].unique()

data['Exterior2nd'].unique()

### Encoding important

We select which digit to assign to each category using the train set, and then use those mappings in the test set.

X_train, X_test, y_train, y_test = train_test_split(
    data[['Neighborhood', 'Exterior1st', 'Exterior2nd']], # predictors
    data['SalePrice'],  # target
    test_size=0.3,  # percentage of obs in test set
    random_state=0)  # seed to ensure reproducibility
X_train.shape, X_test.shape

## Integer encoding with pandas

**It is quick and returns pandas dataframe** But it does not preserve information from train data to propagate to test data! **We need to capture and save the mappings one by one manually**, if we are planing to use those in production.

**Create a dictionary with the mappings of categories to numbers!**

ordinal_mapping = {k: i for i, k in 
                   enumerate(X_train['Neighborhood'].unique(), 0)}
ordinal_mapping

The dictionary indicates which number will replace each category. Numbers were assigned arbitrarily from 0 to n - 1 where n is the number of distinct categories.

**Replace the labels with the integers!**

X_train['Neighborhood'] = X_train['Neighborhood'].map(ordinal_mapping)
X_test['Neighborhood'] = X_test['Neighborhood'].map(ordinal_mapping)

**Explore the result!**

X_train['Neighborhood'].head(10)

**We can turn the previous commands into 2 functions!**

def find_category_mappings(df, variable):
    return {k: i for i, k in enumerate(df[variable].unique(), 0)}

def integer_encode(train, test, variable, ordinal_mapping):
    X_train[variable] = X_train[variable].map(ordinal_mapping)
    X_test[variable] = X_test[variable].map(ordinal_mapping)

**A loop over the remaining categorical variables!**

for variable in ['Exterior1st', 'Exterior2nd']:
    mappings = find_category_mappings(X_train, variable)
    integer_encode(X_train, X_test, variable, mappings)

**The result!**

X_train.head()

## Integer Encoding with Scikit-learn

X_train, X_test, y_train, y_test = train_test_split(
    data[['Neighborhood', 'Exterior1st', 'Exterior2nd']], # predictors
    data['SalePrice'],  # target
    test_size=0.3,  # percentage of obs in test set
    random_state=0)  # seed to ensure reproducibility
X_train.shape, X_test.shape

**Create an encoder!**

le = LabelEncoder()
le.fit(X_train['Neighborhood'])

**See the unique classes!**

le.classes_

X_train['Neighborhood'] = le.transform(X_train['Neighborhood'])
X_test['Neighborhood'] = le.transform(X_test['Neighborhood'])
X_train.head()

Unfortunately, the **LabelEncoder** works one variable at the time. However **there is a way to automate this for all the categorical variables**. I took the below from this [stackoverflow thread](https://stackoverflow.com/questions/24458645/label-encoding-across-multiple-columns-in-scikit-learn)

from collections import defaultdict  # additional import required

X_train, X_test, y_train, y_test = train_test_split(
    data[['Neighborhood', 'Exterior1st', 'Exterior2nd']], # predictors
    data['SalePrice'],  # target
    test_size=0.3,  # percentage of obs in test set
    random_state=0)  # seed to ensure reproducibility
X_train.shape, X_test.shape

d = defaultdict(LabelEncoder)

**Encode the variable! Then use the dictionary to encode future data!**

train_transformed = X_train.apply(lambda x: d[x.name].fit_transform(x))
test_transformed = X_test.apply(lambda x: d[x.name].fit_transform(x))

train_transformed.head()

test_transformed.head()

**Inverse the encoded, to inverse transform to recover the original labels!**

tmp = train_transformed.apply(lambda x: d[x.name].inverse_transform(x))
tmp.head()

Finally, there is another Scikit-learn transformer, the OrdinalEncoder, to encode multiple variables at the same time. However, this transformer returns a NumPy array without column names, so it is not my favourite implementation. More details here: https://scikit-learn.org/stable/modules/generated/sklearn.preprocessing.OrdinalEncoder.html 

## Integer Encoding with Feature-Engine

data1= data[['Neighborhood', 'Exterior1st', 'Exterior2nd', 'SalePrice']].dropna()

X_train, X_test, y_train, y_test = train_test_split(
    data1[['Neighborhood', 'Exterior1st', 'Exterior2nd']], # predictors
    data1['SalePrice'],  # target
    test_size=0.3,  # percentage of obs in test set
    random_state=0)  # seed to ensure reproducibility
X_train.shape, X_test.shape

ordinal_enc = OrdinalEncoder(
    encoding_method='arbitrary',
    variables=['Neighborhood', 'Exterior1st', 'Exterior2nd'])
ordinal_enc.fit(X_train)

**In the encoder dict we can observe the numbers assigned to each category for all the indicated variables!**

ordinal_enc.encoder_dict_

**The list of variables that the encoder will transform!**

ordinal_enc.variables_

**Explore the result!**

X_train = ordinal_enc.transform(X_train)
X_test = ordinal_enc.transform(X_test)
X_train.head()

**Note**

If the argument variables is left to None, then the encoder will automatically identify all categorical variables. Is that not sweet?

The encoder will not encode numerical variables. So if some of your numerical variables are in fact categories, you will need to re-cast them as object before using the encoder.

Note, if there is a variable in the test set, for which the encoder doesn't have a number to assigned (the category was not seen in the train set), the encoder will return an error.
