==> Machine Learning/FEML_Codes/06.09-Comparison-categorical-encoding-techniques.ipynb <==
## Comparison of Categorical Variable Encodings

Now let's compare **the performance** of the different feature categorical encoding techniques: **One hot encoding, Replacing labels by the count, Ordering labels according to target, Mean Encoding and WoE!**

import pandas as pd
import numpy as np
from sklearn.model_selection import train_test_split
from sklearn.linear_model import LogisticRegression
from sklearn.ensemble import RandomForestClassifier
from sklearn.metrics import roc_auc_score

cols = ['pclass', 'age', 'sibsp', 'parch', 'fare',
        'sex', 'cabin', 'embarked', 'survived']
data = pd.read_csv('titanic.csv', usecols=cols)
data.head()

**Check for missing data!**

data.isnull().sum()

**Drop observations with NA in Fare and embarked!**

data.dropna(subset=['fare', 'embarked'], inplace=True)

**Extract the first letter of the cabin!**

data['cabin'] = data['cabin'].astype(str).str[0]
data.head()

**Drop observations with cabin = T, they are too few !**

data = data[data['cabin'] != 'T']

**Divide into train and test set!**

X_train, X_test, y_train, y_test = train_test_split(
    data.drop(labels='survived', axis=1),  # predictors
    data['survived'],  # target
    test_size=0.3,
    random_state=0)
X_train.shape, X_test.shape

**Replace null values in numerical variables by the mean! Impute first the test set!**

def impute_na(df, variable, value):
    df[variable].fillna(value, inplace=True)
impute_na(X_test, 'age', X_train['age'].mean())
impute_na(X_train, 'age',  X_train['age'].mean())

X_train.head()

**Check that we have no missing data after NA imputation!**

X_train.isnull().sum(), X_test.isnull().sum()

### One Hot Encoding

def get_OHE(df):
    df_OHE = pd.concat(
        [df[['pclass', 'age', 'sibsp', 'parch', 'fare']],
         pd.get_dummies(df[['sex', 'cabin', 'embarked']], drop_first=True)], axis=1)
    return df_OHE
X_train_OHE = get_OHE(X_train)
X_test_OHE = get_OHE(X_test)
X_train_OHE.head()

X_test_OHE.head()

### Count encoding

**Make a temporary copy of the original dataframes! Make dictionary mapping category to counts! Remap the labels to their counts!..**

def categorical_to_counts(df_train, df_test):
    df_train_temp = df_train.copy()
    df_test_temp = df_test.copy()
    for col in ['sex', 'cabin', 'embarked']:
        counts_map = df_train_temp[col].value_counts().to_dict()
        df_train_temp[col] = df_train_temp[col].map(counts_map)
        df_test_temp[col] = df_test_temp[col].map(counts_map)

    return df_train_temp, df_test_temp
X_train_count, X_test_count = categorical_to_counts(X_train, X_test)
X_train_count.head()

### Ordered Integer Encoding

**Function: Make a temporary copy of the datasets! Order categories according to target mean! Create the dictionary to map the ordered labels to an ordinal number! Remap the categories  to these ordinal numbers! Then remove the target!**

def categories_to_ordered(df_train, df_test, y_train, y_test):
    df_train_temp = pd.concat([df_train, y_train], axis=1).copy()
    df_test_temp = pd.concat([df_test, y_test], axis=1).copy()
    for col in ['sex', 'cabin', 'embarked']:
        ordered_labels = df_train_temp.groupby(
            [col])['survived'].mean().sort_values().index
        ordinal_label = {k: i for i, k in enumerate(ordered_labels, 0)}
        df_train_temp[col] = df_train[col].map(ordinal_label)
        df_test_temp[col] = df_test[col].map(ordinal_label)
    df_train_temp.drop(['survived'], axis=1, inplace=True)
    df_test_temp.drop(['survived'], axis=1, inplace=True)
    return df_train_temp, df_test_temp
X_train_ordered, X_test_ordered = categories_to_ordered(
    X_train, X_test, y_train, y_test)
X_train_ordered.head()

### Mean Encoding

**Function: Make a temporary copy of the datasets! Calculate mean target per category! Remap the categories to the target mean! Then remove the target!**

def categories_to_mean(df_train, df_test, y_train, y_test):
    df_train_temp = pd.concat([df_train, y_train], axis=1).copy()
    df_test_temp = pd.concat([df_test, y_test], axis=1).copy()
    for col in ['sex', 'cabin', 'embarked']:
        ordered_labels = df_train_temp.groupby(
            [col])['survived'].mean().to_dict()
        df_train_temp[col] = df_train[col].map(ordered_labels)
        df_test_temp[col] = df_test[col].map(ordered_labels)
    df_train_temp.drop(['survived'], axis=1, inplace=True)
    df_test_temp.drop(['survived'], axis=1, inplace=True)
    return df_train_temp, df_test_temp
X_train_mean, X_test_mean = categories_to_mean(
    X_train, X_test, y_train, y_test)
X_train_mean.head()

### Probability Ratio

**Function: Make a temporary copy of the datasets! Create df containing the different parts of the WoE equation prob survived =1 ! Re-map the labels to WoE! Drop the target**

def categories_to_ratio(df_train, df_test, y_train, y_test):
    df_train_temp = pd.concat([df_train, y_train], axis=1).copy()
    df_test_temp = pd.concat([df_test, y_test], axis=1).copy()
    for col in ['sex', 'cabin', 'embarked']:
        prob_df = pd.DataFrame(df_train_temp.groupby([col])['survived'].mean())
        prob_df['died'] = 1-prob_df.survived  # prob survived = 0
        prob_df['Ratio'] = np.log(prob_df.survived/prob_df.died)  # calculate WoE
        woe = prob_df['Ratio'].to_dict()  # capture woe in dictionary
        df_train_temp[col] = df_train[col].map(woe)
        df_test_temp[col] = df_test[col].map(woe)
    df_train_temp.drop(['survived'], axis=1, inplace=True)
    df_test_temp.drop(['survived'], axis=1, inplace=True)
    return df_train_temp, df_test_temp
X_train_ratio, X_test_ratio = categories_to_ratio(X_train, X_test, y_train, y_test)
X_train_ratio.head()

### Random Forest Performance

**Create a function to build random forests and compare performance in train and test set**

def run_randomForests(X_train, X_test, y_train, y_test):
    rf = RandomForestClassifier(n_estimators=50, random_state=39, max_depth=3)
    rf.fit(X_train, y_train)
    print('Train set')
    pred = rf.predict_proba(X_train)
    print(
        'Random Forests roc-auc: {}'.format(roc_auc_score(y_train, pred[:, 1])))
    print('Test set')
    pred = rf.predict_proba(X_test)
    print(
        'Random Forests roc-auc: {}'.format(roc_auc_score(y_test, pred[:, 1])))

**OHE !**

run_randomForests(X_train_OHE, X_test_OHE, y_train, y_test)

**Counts!**

run_randomForests(X_train_count, X_test_count, y_train, y_test)

**Ordered labels!**

run_randomForests(X_train_ordered, X_test_ordered, y_train, y_test)

**Mean encoding!**

run_randomForests(X_train_mean, X_test_mean, y_train, y_test)

**Ratio!**

run_randomForests(X_train_ratio, X_test_ratio, y_train, y_test)

Comparing the roc_auc values on the test sets, we can see that one hot encoding has the worse performance. This makes sense because trees do not perform well in datasets with big feature spaces.

The remaining encodings returned similar performances. This also makes sense, because trees are non-linear models, so target guided encodings may not necessarily improve the model performance

### Logistic Regression Performance

**Function to train and test the performance of logistic regression!**

def run_logistic(X_train, X_test, y_train, y_test):
    logit = LogisticRegression(random_state=44, C=0.01, max_iter=100)
    logit.fit(X_train, y_train)
    print('Train set')
    pred = logit.predict_proba(X_train)
    print(
        'Logistic Regression roc-auc: {}'.format(roc_auc_score(y_train, pred[:, 1])))
    print('Test set')
    pred = logit.predict_proba(X_test)
    print(
        'Logistic Regression roc-auc: {}'.format(roc_auc_score(y_test, pred[:, 1])))

**OHE !**

run_logistic(X_train_OHE, X_test_OHE, y_train, y_test)

**Counts!**

run_logistic(X_train_count, X_test_count, y_train, y_test)

**Ordered labels!**

run_logistic(X_train_ordered, X_test_ordered, y_train, y_test)

**Mean encoding!**

run_logistic(X_train_mean, X_test_mean, y_train, y_test)

**Ratio!**

run_logistic(X_train_ratio, X_test_ratio, y_train, y_test)

For Logistic regression, the best performances are obtained with one hot encoding, as it preserves linear relationships with variables and target, and also with weight of evidence, and ordered encoding.

Note however how count encoding, returns the worse performance as it does not create a monotonic relationship between variables and target, and in this case, mean target encoding is probably causing over-fitting.
