==> Machine Learning/FEML_Codes/09.05-Capping-Arbitrary.ipynb <==
## Outlier Engineering

We will perform **capping with arbitrary values** using the Titanic dataset!

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
from feature_engine.outliers import ArbitraryOutlierCapper

**Function to load the titanic dataset!..**

def load_titanic():
    data = pd.read_csv('../titanic.csv')
    data['cabin'] = data['cabin'].astype(str).str[0]
    data['pclass'] = data['pclass'].astype('O')
    data['embarked'].fillna('C', inplace=True)
    return data

data = load_titanic()
data.head()

## ArbitraryOutlierCapper

The **ArbitraryOutlierCapper** caps the **minimum and maximum values** by a value determined by the user. 

**Find out the maximum Age and maximum Fare in the titanic!**

data.age.max(), data.fare.max()

capper = ArbitraryOutlierCapper(max_capping_dict={'age': 50, 'fare': 200},
                                min_capping_dict=None)
capper.fit(data.fillna(0))

capper.right_tail_caps_

capper.left_tail_caps_

temp = capper.transform(data.fillna(0))
temp.age.max(), temp.fare.max()

### Minimum capping

capper = ArbitraryOutlierCapper(max_capping_dict=None,
                                min_capping_dict={'age': 10, 'fare': 100})
capper.fit(data.fillna(0))

capper.variables_

capper.right_tail_caps_

capper.left_tail_caps_

temp = capper.transform(data.fillna(0))

temp.age.min(), temp.fare.min()

### Both ends capping

capper = ArbitraryOutlierCapper(max_capping_dict={
    'age': 50, 'fare': 200},
    min_capping_dict={
    'age': 10, 'fare': 100})

capper.fit(data.fillna(0))

capper.right_tail_caps_

capper.left_tail_caps_

temp = capper.transform(data.fillna(0))

temp.age.min(), temp.fare.min()

temp.age.max(), temp.fare.max()

## Feature-engine

We can also automate this process with Feature-engine:
https://feature-engine.readthedocs.io/en/latest/outliers/ArbitraryOutlierCapper.html
