==> Machine Learning/FEML_Codes/12.02_Engineering_time.ipynb <==
## Engineering Time

We will extract different ways of representing time from a **timestamp**. We can extract for example: **hour, minute, second, data, elapsed time**! We will create a toy dataset for the demonstration.

import pandas as pd
import numpy as np
import datetime

**Create a toy data set: 1 column 7 different timestamps, 1 hr difference between timestamp!**

date = pd.Series(pd.date_range('2015-1-5 11:20:00', periods=7, freq='H'))
df = pd.DataFrame(dict(date=date))
df

### Extract the hr, minute and second

df['hour'] = df['date'].dt.hour
df['min'] = df['date'].dt.minute
df['sec'] = df['date'].dt.second
df

### Extract time part

df['time'] = df['date'].dt.time
df

### Extract hr, min, sec, at the same time

Repeat what we did in cell 3 in 1 command!

df[['h','m','s']] = pd.DataFrame([(x.hour, x.minute, x.second) for x in df['time']])
df

## Calculate time difference

Create another **toy dataframe** with **2 timestamp columns** and **7 rows** each, in the first column the timestamps change **monthly**, in the second column the timestamps change **weekly**

date1 = pd.Series(pd.date_range('2012-1-1 12:00:00', periods=7, freq='M'))
date2 = pd.Series(pd.date_range('2013-3-11 21:45:00', periods=7, freq='W'))
df = pd.DataFrame(dict(Start_date = date1, End_date = date2))
df

**Calculate the time elapsed in seconds!**

df['diff_seconds'] = df['End_date'] - df['Start_date']
df['diff_seconds']=df['diff_seconds']/np.timedelta64(1,'s')
df

**Calculate the time elapsed in minutes!**

df['diff_seconds'] = df['End_date'] - df['Start_date']
df['diff_seconds']=df['diff_seconds']/np.timedelta64(1,'m')
df

For more details visit [this article](http://www.datasciencemadesimple.com/difference-two-timestamps-seconds-minutes-hours-pandas-python-2/)

## Work with different timezones

In the next few cells, we will see how to work with timestamps that are in different time zones.

**Create a toy dataframe with some timestamps in different time zones!**

df = pd.DataFrame()
df['time'] = pd.concat([
    pd.Series(
        pd.date_range(
            start='2014-08-01 09:00', freq='H', periods=3,
            tz='Europe/Berlin')),
    pd.Series(
        pd.date_range(
            start='2014-08-01 09:00', freq='H', periods=3, tz='US/Central'))
    ], axis=0)
df

We can see the different timezones indicated by the +2 and -5, respect to the meridian.

**To work with different time zones, first we unify the timezone to the central one setting utc = True! Next we change all timestamps to the desired timezone, eg Europe/London!**

df['time_utc'] = pd.to_datetime(df['time'], utc=True)
df['time_london'] = df['time_utc'].dt.tz_convert('Europe/London')
df

For **feature engineering**, we can of course, set all timezones to the central, utc=True, and work with that to extract time elapsed etc. **The additional timezone encoding is mostly for human readability**.
